front/requirements.txt
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
opencv-python==4.5.4.60
Pillow==8.4.0
pydantic==1.8.2
requests==2.26.0
Shapely==1.8.0
urllib3==1.26.7
Jinja2==3.0.3
streamlit==1.3.0